refactor: move offerings namespace into issuance namespace
Also expose issuance namespace (wasn't being exposed properly before)

BREAKING CHANGE: `securityToken.offerings` is now `securityToken.issuance.offerings`
