+++ src/entities/SecurityToken/Issuance/index.ts
@@ -1,0 +1,1 @@
+export { Issuance } from './Issuance';
