+++ src/entities/SecurityToken/SecurityToken.ts
@@ -4,12 +4,12 @@ import { serialize, unserialize as unserializeUtil } from '../../utils';
 import { Features } from './Features';
 import { Shareholders } from './Shareholders';
 import { Dividends } from './Dividends';
-import { Offerings } from './Offerings';
+import { Issuance } from './Issuance';
 import { Permissions } from './Permissions';
 import { Transfers } from './Transfers';
+import { Documents } from './Documents';
 import { PolymathError } from '../../PolymathError';
 import { ErrorCode } from '../../types';
-import { Documents } from './Documents';
 
 export interface UniqueIdentifiers {
   symbol: string;
@@ -65,7 +65,7 @@ export class SecurityToken extends Entity<Params> {
 
   public dividends: Dividends;
 
-  public offerings: Offerings;
+  public issuance: Issuance;
 
   public permissions: Permissions;
 
@@ -86,7 +86,7 @@ export class SecurityToken extends Entity<Params> {
     this.features = new Features(this, context);
     this.shareholders = new Shareholders(this, context);
     this.dividends = new Dividends(this, context);
-    this.offerings = new Offerings(this, context);
+    this.issuance = new Issuance(this, context);
     this.permissions = new Permissions(this, context);
     this.transfers = new Transfers(this, context);
     this.documents = new Documents(this, context);
